refactor: rename to convertPlaceToCoordinate
