+++ server/src/routes/api/events/controllers/convertPlaceToCoordinate.ts
