+++ server/src/routes/api/events/controllers/index.ts
@@ -1,3 +1,3 @@
 export { default as getEvents } from './getEvents';
 export { default as getEvent } from './getEvent';
-export { default as placeToCoordinate } from './placeToCoordinate';
+export { default as convertPlaceToCoordinate } from './convertPlaceToCoordinate';
